+++ src/classifier/index.ts
@@ -1,0 +1,4 @@
+export { EMAIL_KINDS } from "./types.js";
+export type { EmailKind, EmailClassification, EmailClassifier } from "./types.js";
+export { RulesClassifier } from "./rules.js";
+export { JevClassifier, JevClassifierError, type JevClassifierConfig } from "./jev.js";